Added admin panel, recipes and nutrition

+++ internal/handler/activity.go
@@ -22,7 +22,7 @@ type CreateActivityRequest struct {
 	Distance  *float64  `json:"distance,omitempty"`
 	Duration  *int      `json:"duration,omitempty"`
 	Elevation *float64  `json:"elevation,omitempty"`
-	Calories  *float64  `json:"calories,omitempty"`
+	Calories  *int      `json:"calories,omitempty"`
 	StartedAt time.Time `json:"started_at" binding:"required"`
 	Source    string    `json:"source,omitempty"`
 	SourceID  string    `json:"source_id" binding:"required"`
